move ceremonyBoxService.dart to ceremonyBox
