+++ lib/page-encointer/ceremony_box/ceremonyProgressBar.dart
@@ -1,5 +1,5 @@
 import 'package:encointer_wallet/common/theme.dart';
-import 'package:encointer_wallet/service/ceremonyBoxService.dart';
+import 'package:encointer_wallet/page-encointer/ceremony_box/ceremonyBoxService.dart';
 import 'package:encointer_wallet/store/encointer/types/ceremonies.dart';
 import 'package:flutter/material.dart';
