+++ lib/page-encointer/ceremony_box/ceremonySchedule.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:iconsax/iconsax.dart';
 import 'package:intl/intl.dart';
 import 'package:encointer_wallet/common/theme.dart';
-import 'package:encointer_wallet/service/ceremonyBoxService.dart';
+import 'package:encointer_wallet/page-encointer/ceremony_box/ceremonyBoxService.dart';
 import 'package:encointer_wallet/utils/translations/index.dart';
 
 import 'ceremonyCountDown.dart';
